Refactor schema import paths for Agent and Message
