+++ apps/adk-web/app/(dashboard)/_components/agent-card.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@/components/ui/button";
 import { cn } from "@/lib/utils";
 import { Play, Square } from "lucide-react";
-import type { Agent } from "../_schema/types";
+import type { Agent } from "../_schema";
 
 interface AgentCardProps {
 	agent: Agent;
