+++ apps/adk-web/components/chat-panel.tsx
@@ -1,4 +1,4 @@
-import type { Agent, Message } from "@/app/(dashboard)/_schema/types";
+import type { Agent, Message } from "@/app/(dashboard)/_schema";
 import { Avatar, AvatarFallback } from "@/components/ui/avatar";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
